Move CountryChip component to components/UI.
